[ SFT REVISION ]minor, FILE BSP API attribut type change

+++ src/bsp/platform/posix/xi_bsp_io_fs_posix.c
@@ -128,7 +128,7 @@ err_handling:
 }
 
 xi_state_t xi_bsp_io_fs_open( const char* const resource_name,
-                              const uint32_t size,
+                              const size_t size,
                               const xi_fs_open_flags_t open_flags,
                               xi_fs_resource_handle_t* resource_handle_out )
 {
